Serverside data/checks for unit spawning. Refactoring.
Server now does (more) of the calculations for unit spawning, and doesn't rely on the client to get important values like spawn time and cost. Also began refactoring this part of the code in general.

The impact on networking should be less as we're now sending substantially less data (before we were sending entire tables and strings, now just a few ints/bools and 2 vectors). This should in theory make glitches less likely to occur, and a bunch of other networking-related issues less likely to happen.

Optimally I want to make the checks for if a unit can spawn into a shared function that both server and client call, so client can determine display values, but server doesn't rely on the client to give it reliable data. This'd be a lot better than the structure we have right now which is incredibly messy.