Shared unit-spawning checks.
All checks performed to determine whether you're allowed to spawn a unit or not are now performed in a single function, which is called both on client and server. There were a large number of checks that were on either server/client but not the other, and the entire system was overall messy. This should reduce the frequency of server-client de-syncs to almost nothing, and should make weird rare issues like not having spawn time/being able to go in-debt impossible.