Misc unit optimizations and bugfixes.
improved the clientside draw and think functions of units. Also formatted the file. Almost all changes are either reducing repeated work, or reducing indexing of the entity. Notably I did entirely remove something made to debug selection bug (for a moderate performance boost). I don't know when this was added, but it must have been in vanilla. In-spite of this I have literally never seen it actually do anything, so I think it's safe to say it doesn't work and hence isn't worth keeping.

I also optimized a few of the frequently-running serverside functions for units. calls to cvars.Bool were replaced with :GetBool() on a stored convar object, as this is less expensive.
DefaultPhysicsUpdate had its repeat work slightly reduced, and the number of new vector objects being created reduced considerably.

unitDefaultThink had the largest number of changes, with its loops being made more efficient. I also fixed some of the visibility-detection code not actually doing anything. Which was resulting in melons on the other side of walls "distracting" units. They wouldn't be shot at, but would be targeted (often instead of valid targets, which was an issue).