Improved unit spawn menus, bugfixes
Fixed misnamed variables causing energy radius indicators to break

Standardized the unit-lists (energy, buildings, units) to use a single function to generate, instead of having large amounts of repeated code in each. Removed the broken selection-indicator-thing and replaced it with a simpler (and prettier) solution, that being darkening the buttons themselves when hovered.