Reduced per-unit NW variable count from 8 to 6
Removed range NW variable as it was only used for the secret unit-controlling mode. Also commented out unit-control code for now. I'll see if I can bring it back and make it so that range is only set when you start controlling a unit. That's a kinda hacky solution (it'd be better to just network it to the client a single time and not use NW vars at all) but it's the laziest/easiest I can think of to implement, and I don't care much because this is a niche and completely hidden feature anyway.

Replaced health and maxhealth NWVars with a single healthFraction NW var, as this was the only (valid) use for them. A bunch of units used the NWVar instead of the actual HP variable, even though they were all serverside. This has been fixed. I've probably just introduced a bunch of bugs/lua errors, but I should be able to find any important ones through playtesting. And it's *probably* worth it for the significantly(?) reduced networking.