Rewrote/restructured the contraption networking code.
Dedicated clientside function for handling sending contraptions (which may be too large for a single net message) to the server.

Separated loaded and spawning logic, so that now we have one spawning function that we call as needed, rather than needing weird chains of network messages between client server (e.g. spawning via assembler used to start with a request from client, then later the assembler would request the client to send its data to the server, and then the server would spawn it with that.)

net messages following the load message now determine what to do with the loaded contraption (either loading it to an assembler, or trying to spawn it).

The new structure will in the future allow me to implement proper legality checks, as we are no longer relying entirely on the client to tell us what the contraption's stats are (without actually having the contraption data itself until we spawn it). I have not started on this yet however.

Assemblers also now actually work. I fixed a logic mistake craft (who is probably the only one who will read this, aside from me) made which was causing their cost checks to fail when infinite water was on.