range indicators for buildings, further tool organisation
Re-organized the tool file again, splitting the rendering/HUD-related functions off from "Tool Functions", and re-ordering Tool Functions so that related functions are adjacent (i.e. deploy next to holster, rightclick next to leftclick, etc)

Added a world-rings-based range indicator to the spawner tool, which displays attack range for buildings, utility-effect range, and how close you need units for the few buildings that can only be built near friendlies. I've added this to every building I could find that has some sort of range.

This idea was based on/inspired by a feature request I got a long while back, where someone asked me to make the turret's range visible while spawning. I did actually do this (that's the reason the turret has had an energy range marker all this time, it was the easiest way to do that.) but it was a very hacky/low effort solution, largely because I didn't really have the time, knowledge, or inclination to make a better one. WorldRings presented an opportunity because they're very well-suited to this task, and would be relatively easy to add. So I did that.

I also moved just a little bit of the clientside contraption logic/calculations over to server. I'm of course far from done with that still.