Better actuate, better hovers
Actuate no longer relies on a list of hard-coded actuatable entities, which must be checked on the client. Instead the client simply sends an entity and the server actuates it if it can be actuated.

BarracksToggle and hover/propeller activation now make use of the actuate function instead of having their own dedicated net messages.