Editable Melonium-Plate Model
Not much more to add that isn't in the title already.

Previously I relied on model-manipulator to let people change the model of this entity. That's another "Fine for a small community, not fine for the workshop" sort of thing. I can't expect nearly that amount of ingenuity from the general public, so this is a solution that should work well-enough for at least the people who are invested enough to use it in their prop maps.

Also I hate editable entities. They're so jank to work with and every time I use them I feel like I'm committing some sort of programming sin with the things I have to do to get them working how I want them.