Improved ParticleBeam Explosion FX, some work on contrap validation, partially cleaned up unit launchers
Replaced particle beam's old explosion effect (which was just the nuke's explosion) with a fancier new one. (Bright flash/combine energy pulse on the ground).

Server now checks the mass values for props instead of trusting the file's cost data, meaning contraption cost and spawntime values are now 100% resistant to file-editing, as costs and spawn times are calculated when the contraption is loaded. Individual unit stats are still unvalidated and will need more widespread changes (or a new method) to fix.

Did a decent bit of work cleaning up unit launcher code. I still have a fair bit to do but this is already a lot better than it was. I think I need to do largely the same thing to transports as most of this code was copied from them.
Also added a range indicator for it.

Range indicators now scale their alpha for rings with large radii. They become more opaque linearly above 800 radius. This change was made as the rings for long-range units like particle towers and unit launchers were incredibly difficult to see.