Fighters work, jetpack refactor.
Fighters now actually work and are usable. Their code kinda fucking sucks but at least they can be played with now.

Minor optimization/refactor for jetpacks. Just carried over some of the same stuff that I did for propellers.

Refactor for stop order. Both to make it cleaner/more optimized, and to enable me to do more with it (this was needed to make fighters work).