Molotov Barracks model fix, HQ refactor
While checking to make sure all barracks still produced the correct units, I realized molotovs used a non-existent model. I believe this is an episodic model, as I remember seeing it previously and I'm pretty sure this is the same model used in vanilla MW. I un-mounted my game content while working on a previous update to figure out which particles would be available, this issue was probably found as a result of that.

Cleaned up the HQ files a little bit. I forgot to do ornaments, but I'm going to bed now so I'll do that tomorrow.