Ornament refactor & ent_melon_base optimization
Did the same to ornaments that I did to the other HQs

Applied some really hacky optimizations to ent_melon_base's physics collide. Somewhat questionable as to whether it was worth it or not, because the difference is incredibly minor.

Applied some considerably less hacky optimizations to update as sort of an afterthought and managed a far greater performance improvement (lol).