Polished buildspeed modifiers, General refactoring
Renamed buildspeed modifiers to buildspeed multipliers to make their purpose a bit clearer. Changed references to the actual applied value to call it a spawn-time multiplier, as that's more accurate to what it does mechanically. Having 2 different names is maybe not the best but hopefully this should make it both clear *at a glance* what the general purpose of it is, and for those using it make it obvious what it actually changes and how.

Made the messages it prints clearer / nicer, and removed the on-spawn one. Added an E in the style of water tanks/hovers/propellers to indicate to players how to activate it.

Made water tank E indicator only display when it's fully captured.

Some minor optimization for ent_melon_base