Refactoring, cleaned up some TODOs
Removed un-used code from SpawnUnitAtPos
moved spawn sound effect out of SpawnUnitAtPos

Made a function to handle sending credit values to all members of a particular team, as it was repeated in a lot of places.

Rounded the print value on spawntimemodifiers

moved wheel update to wheel think, as it doesn't need any of the vanilla units' think logic.

Added blockfriendlytraces value on entities to get rid of some hardcoding.

Fucked with engines a bit but didn't finish working on them.

Partial refactor of BarracksSlowThink