muzzleOffset, transport refactor, engineer/medic refactor, admin code
Refactored transport. Not much to say here, just copied/reworked the code from unit launchers.

Refactored engineer and medic. They both use the generic find units function now. They still have a lot of duplicate code but this is good enough for now. longterm it'd be better to have it handled by a function like many other things are.

Made admin code units only show up if you have the right code *and* are an admin. This check only happens clientside so it's not completely secure, but it makes it harder to abuse, enough so that no cheater I've encountered would have the technical knowledge to do so.

Separated the shot offset for shooting from the shot offset for being shot. Currently this is only used for turrets.