Minor cleanup / bug fixes
Fixed unit transport being constantly set to sleep due to it being immobile.

Swapped a few hardcoded class checks with unit.isContraptionPart.

Refactored engine files. I wanted to remove the custom updates entirely and replace them with just generic logic, but that doesn't seem like it'll be possible. At least not easily. I optimized/cleaned them up a bit but that's going to be all for now, I want to wrap this up and I don't think doing more is worth the time.