Fixed openpermits & made every admin-option functional on dedicated
Fixed setConvar net message returning if openPermits was true, rather than when it was false, causing it to not allow any changes.

Made the other netmessages (sandbox, start game, base spawning) also check if a player is allowed to use them. This wasn't a big deal but did mean a technically experienced cheater could potentially use some admin-menu options without being an admin or having access to the menu.

Added a duplicate of the SetMWConvar net message, which handles integers instead, and made the resource sliders call it. This means team-income/starting water/etc can be set now.

Made sandbox mode set convars serverside as well as clientside, allowing it to work on dedicated servers.