Really hacky entity damage-protection work-around.
If a CPPI compliant prop-protect addon is enabled, a hook is added to bypass damage protection on entities using ent_melon_base (or any other melon-baseclass). This is done by setting the attacker/inflictor to be the melon receiving damage, which should make it reliably pass permissions-checks implemented by prop-protection addons. The entity's OnTakeDamage is called first before doing this, and the next call suppressed, as otherwise the target's damage multipliers would be applied instead of the attacker's.