Cleared a few more TODOs
Removed duplicate / outdated damagedeal logic in tesla

Refactored shredder and made it more resistant to potential exploits.

Mild refactor/optimization of gatlings.

LoadToAssembler now checks if we have the resources to build a contraption before allowing us to do so (or charging us).

Removed or asterisk-ed the TODOs in ent_melon_energy_base