Finished off the TODOs. Set units/Baseprops CPPIOwner
Culled/rewrote the about section to be substantially more concise. I removed most of the "why" and just kept the what.

Filled-in the factions section, including basic descriptions of each faction and some important information on how to handle them for those who can't figure things out by themselves.

Did a quick pass on the bonus units to ensure there weren't any glaring issues with them. They're going to be a lot more error-prone and unoptimized than everything else, but that's not a priority to fix given they're not the intended way to play the game.

isBarracks value for barracks, overclocker uses this.

Uplink refactor (even though it's a secret unit. And this probably wasn't important to do.)

Improved barracksSlowthink a bit more. It's still not great but I don't really care to make it better.

base prop and unit spawn functions set CPPI owner if prop-protection is present. This isn't that important but I guess it'll prevent some cheating on servers (I've seen people physgun contraptions or context-menu remove units before. Both blatantly and covertly. So this would prevent that.)