Fixed contraption system on 64bit
Player being passed to the duplicator paste function was null. This was breaking things on the 64 bit beta branch, as the glua duplicator code for that is different and assumes the player exists.

This is the issue I was trying to fix in my previous commit. I didn't have enough information at the time to reproduce it.