Really hacky fix to broken contraptions
At some point an IsSpawnable check got added to gmod's duplicator code. This is really annoying because this is only ever called serverside with pre-validated data.

I now register every melon wars entity with a factory function, and use a modified (copy-pasted with parts removed) version of the default factory function.

The default factory function is actually accessible, but the isAllowed check is hardcoded into it, and all of the functions it calls are inaccessible. Which is again, *very annoying*.