More descriptive error if tool file doesn't load, minor optimisation, updated outdated terminology
MW:F Now produces a console error telling people to disable the vanilla version if the tool-file fails to load. Hopefully this will save people from needing to go to the issue reports section to ask about it, as that has probably been the most frequent cause of issue reports.

Changed convar-caching to use AddChangeCallback, like it should have in the first place. This will improve performance, but only by a miniscule amount.

Updated tutorial text to say engine instead of thruster, to avoid confusing people.